Make ContentItem presenter generic
Content item presenter was used to render the CaseStudy format and had a lot of
logic that is specific to this format. In order to make ContentItemPresenter
useful for rendering other content types, keep only generic functionality in
there and extract CaseStudy-specific logic to a new CaseStudyPresenter class.
