+++ app/controllers/content_items_controller.rb
@@ -27,7 +27,7 @@ private
     end
 
     case content_item['format']
-      when 'case_study' then ContentItemPresenter.new(content_item)
+      when 'case_study' then CaseStudyPresenter.new(content_item)
       when 'unpublishing' then UnpublishingPresenter.new(content_item)
       when 'coming_soon' then ComingSoonPresenter.new(content_item)
       else raise "No support for format \"#{content_item['format']}\""
